requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
distlib==0.3.4
filelock==3.4.2
frozenlist==1.3.0
gevent==21.12.0
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
greenlet==1.1.2
gspread==5.1.1
httplib2==0.20.4
idna==3.3
multidict==6.0.2
numpy==1.22.2
oauth2client==4.1.3
oauthlib==3.2.0
pandas==1.4.0
platformdirs==2.5.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.13.1
yarl==1.7.2
zope.event==4.5.0
zope.interface==5.4.0
